+++ test_packages/config_test/test/values.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+// Colors
+const testColor01 = Color(0xff000001);
+const testColor02 = Color(0xff000002);
+const testColor03 = Color(0xff000003);
+const testColor04 = Color(0xff000004);
+const testColor05 = Color(0xff000005);
+const testColor06 = Color(0xff000006);
+const testColor07 = Color(0xff000007);
+const testColor08 = Color(0xff000008);
+const testColor09 = Color(0xff000009);
+const testColor10 = Color(0xff000010);
+
+// Icons
+const testIcons01 = "test_assets/test_folder/icons_folder_test/example_test/aliqua.svg";
+const testIcons02 = "test_assets/test_folder/icons_folder_test/example_test/eiusmod-dolore.svg";
+const testIcons03 = "test_assets/test_folder/icons_folder_test/example_test/eiusmod-tempor-dark_test.svg";
+const testIcons04 = "test_assets/test_folder/icons_folder_test/example_test/eiusmod-tempor-light_test.svg";