+++ test_packages/general_test/test/token_strings_test.dart
@@ -1,0 +1,18 @@
+import 'package:general_test/ui/ui.ui.dart';
+import 'package:test/test.dart';
+
+import 'values.dart';
+
+void main() {
+  test('Token: numbers', () {
+    expect(StringsXml.lorem, testString01);
+    expect(StringsXml.ipsumDolor, testString02);
+    expect(StringsXml.sitAmet, testString03);
+    expect(StringsXml.consecteturAdipiscingElit, testString04);
+    expect(StringsXml.sedDoEiusmodTemporIncididunt, testString05);
+    expect(StringsXml.utLaboreEtDolore, testString06);
+    expect(StringsXml.magnaAliqua, testString07);
+    expect(StringsXml.utEnimAdMinimVeniam, testString08);
+    expect(StringsXml.quisNostrudExercitationUllamco, testString09);
+  });
+}